station3/requirements.txt
attrs==22.2.0
av==12.3.0
blinker==1.9.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.2.1
colorzero==2.0
contourpy==1.3.3
cycler==0.12.1
distro==1.8.0
Flask==3.1.1
fonttools==4.59.0
gpiozero==2.0.1
idna==3.3
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpointer==2.3
jsonschema==4.10.3
kiwisolver==1.4.8
lgpio==0.2.2.0
libarchive-c==2.9
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
numpy==1.24.2
olefile==0.46
packaging==25.0
picamera2==0.3.30
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==9.4.0
pycryptodomex==3.11.0
PyOpenGL==3.1.6
pyparsing==3.2.3
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
python-apt==2.6.0
python-dateutil==2.9.0.post0
python-prctl==1.8.1
requests==2.28.1
rfc3987==1.3.8
rpi-lgpio==0.6
simplejpeg==1.8.1
six==1.16.0
smbus2==0.4.2
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.12
videodev2==0.0.4
webcolors==1.11.1
Werkzeug==3.1.3
